ENH: Update pre-commit and repository
