+++ src/initial_implementation/AlternativeFunctions.py
