+++ tests/test_vtk_helpers.py
@@ -1,5 +1,6 @@
 """PyDRex: Tests for VTK readers and helpers."""
 import numpy as np
+
 from pydrex import vtk_helpers as _vtk
